Alter the PointcutInterface::matchesMethod to include the ClassReflection instance
The current PointcutInterface methods signature is only valid if one Pointcut
handle one class. However if a Pointcut is implemented to handle more than
one class then the model is broken. The MethodReflection contains the class
where the method is defined and not the class referenced in a sub type class.

So this patch solves the issue by providing the ReflectionClass targetted by
the configuration.